:tada: Initial restructuring and refactoring (WIP)

+++ src/Slip39/BitStreamWriter.cs
@@ -4,11 +4,10 @@ class BitStreamWriter(BitStream stream)
 {
     public BitStreamWriter()
         : this(new BitStream(new byte[100]))
-    { }
+    {
+    }
 
-    public void Write(long data, int count) =>
-        stream.WriteBits(data, (byte)count);
+    public void Write(long data, int count) => stream.WriteBits(data, count);
 
-    public byte[] ToByteArray() =>
-        stream.ToByteArray();
+    public byte[] ToByteArray() => stream.ToByteArray();
 }
